Factor-out reductions from lax_numpy.py
